Make Native and CUDA pooling states public
