+++ chainerx_cc/chainerx/native/native_device.h
@@ -3,21 +3,68 @@
 #include <cstdint>
 #include <memory>
 #include <tuple>
+#include <utility>
 
 #include <nonstd/optional.hpp>
 
 #include "chainerx/array.h"
 #include "chainerx/axes.h"
 #include "chainerx/device.h"
+#include "chainerx/dtype.h"
 #include "chainerx/indexable_array.h"
 #include "chainerx/indexer.h"
 #include "chainerx/native/native_backend.h"
 #include "chainerx/routines/pooling.h"
 #include "chainerx/scalar.h"
+#include "chainerx/shape.h"
 #include "chainerx/stack_vector.h"
 
 namespace chainerx {
 namespace native {
+
+class NativeMaxPoolGradState : public MaxPoolGradState {
+public:
+    NativeMaxPoolGradState(Array x, Array col, Axes axes) : x_{std::move(x)}, col_{std::move(col)}, axes_{std::move(axes)} {}
+
+    const Array& x() const { return x_; }
+    const Array& col() const { return col_; }
+    const Axes& axes() const { return axes_; }
+
+private:
+    Array x_{};
+    Array col_{};
+    Axes axes_{};
+};
+
+class NativeMaxPoolGradGradState : public MaxPoolGradGradState {
+public:
+    NativeMaxPoolGradGradState(Array indices, Array offset, Dtype x_dtype)
+        : indices_{std::move(indices)}, offset_{std::move(offset)}, x_dtype_{x_dtype} {}
+
+    const Array& indices() const { return indices_; }
+    const Array& offset() const { return offset_; }
+    Dtype x_dtype() const { return x_dtype_; }
+
+private:
+    Array indices_{};
+    Array offset_{};
+    Dtype x_dtype_{};
+};
+
+class NativeAveragePoolGradState : public AveragePoolGradState {
+public:
+    NativeAveragePoolGradState(Array x, Shape gcol_shape, nonstd::optional<Array> width_ignore)
+        : x_{std::move(x)}, gcol_shape_{std::move(gcol_shape)}, width_ignore_{std::move(width_ignore)} {}
+
+    const Array& x() const { return x_; }
+    const Shape& gcol_shape() const { return gcol_shape_; }
+    const nonstd::optional<Array>& width_ignore() const { return width_ignore_; }
+
+private:
+    Array x_;
+    Shape gcol_shape_;
+    nonstd::optional<Array> width_ignore_;
+};
 
 class NativeDevice : public Device {
 public:
